+++ CHANGELOG.md
@@ -1,4 +1,23 @@
+# Changelog
+
 ## [Unreleased]
+
+## [0.18.0] - 2022-01-14
+
+### Added
+
+-   ReadTheDocs documentation in `docs` directory.
+-   `{{ var.value.VARIABLE_NAME }}` gets replaced with Airflow variables when parsing `airflow.yml` file.
+
+### Changed
+
+-   Rename project from `dbt-airflow-manifest-parser` to `dbt-airflow-factory`.
+
+### Fixed
+
+-   `KubernetesExecutionParameters.env_vars` works in Airflow 1 too. Airflow 1 is expecting a real dictionary of
+    environment variables instead of a list of `k8s.V1EnvVar` objects.
+-   Fix `DummyOperator` import in `operator.py` to work in Airflow 1.
 
 ## [0.17.0] - 2022-01-11
 
@@ -56,9 +75,11 @@
 
 ## [0.2.0] - 2021-10-25
 
--   Initial implementation of `dbt_airflow_manifest_parser` library. 
+-   Initial implementation of `dbt_airflow_manifest_parser` library.
 
-[Unreleased]: https://github.com/getindata/dbt-airflow-manifest-parser/compare/0.17.0...HEAD
+[Unreleased]: https://github.com/getindata/dbt-airflow-factory/compare/0.18.0...HEAD
+
+[0.18.0]: https://github.com/getindata/dbt-airflow-factory/compare/0.17.0...0.18.0
 
 [0.17.0]: https://github.com/getindata/dbt-airflow-manifest-parser/compare/0.16.0...0.17.0